refactor(parser): use parseModuleFixture from the new parsers in other files
